Move type definitions for transfer protocol from client to transfer
